Skip first two transaction: bootstrap and user schema.
It's technically very challenging to consider syncing Mentat stores
with different bootstrap transactions.  Therefore, Tolstoy blindly
skips the first (bootstrap) transaction.  In the future, we can at
least recognize that the underlying Mentat stores are incompatible and
bail out early.

Additionally, since merge is currently not supported, it's necessary to
also skip the "first" user schema transaction - it's assumed that applicatons
define their schema as the first order of business, and that starting schemas
are the same for a given application that wants to sync.

Both of these "hacks" are temporary, the latter particularly so.